Simplify usage of Executor
Now executor does not need to know the size of the stage in advance.
Instead of immediately dispatching submitted tasks to the thread pool,
tasks are accumulated in the internal queue, and run only after calling
runStage(). Then the barrier of appropriate size (equal to number of
previously submitted tasks) is created, tasks are executed and the call
returns only after the stage has finished. Such design significantly
simplifies usage of the executor, since there is no longer need to
calculate number of productions in the stage.